SAMPLE: Changed default input for loop to int.

+++ SAMPLE/Main.java
@@ -5,15 +5,15 @@ public class Main {
 	private static final PrintStream out = new PrintStream(new BufferedOutputStream(System.out));
 	private static final BufferedReader in = new BufferedReader(new InputStreamReader(System.in));
 
-	public static void output(long x, long y) throws Exception {
+	public static void output(int x, int y) throws Exception {
 	}
 
 	public static void main(String args[]) throws Exception {		
 		int testCases = Integer.parseInt(in.readLine());
-		for(long i = 0; i < testCases; i++) {
+		for(int i = 0; i < testCases; i++) {
 			StringTokenizer tokenizer = new StringTokenizer(in.readLine());
-			long x = Long.parseLong(tokenizer.nextToken()),
-				 y = Long.parseLong(tokenizer.nextToken());
+			int x = Integer.parseInt(tokenizer.nextToken()),
+				y = Integer.parseInt(tokenizer.nextToken());
 
 			output(x, y);
 			out.println();
